pyproject.toml
[tool.poetry]
name = "asciidump"
version = "0.0.1"
description = ""
authors = ["Matthew Barber <quitesimplymatt@gmail.com>"]
license = "BSD-3-Clause"

[tool.poetry.scripts]
asciidump = 'asciidump:main'

[tool.poetry.dependencies]
python = "^3.8"
Jinja2 = "^3.0.1"
rich = "^10.11.0"
libsass = "^0.21.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
libsass = "^0.21.0"
pre-commit = "^2.15.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
